+++ test/cases/option-extract-comments-false/webpack.config.js
@@ -1,0 +1,48 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+  //mode: 'development',
+
+  entry: {
+    index: './src/index.pug',
+  },
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+    filename: 'assets/js/[name].[contenthash:8].js',
+    clean: true,
+  },
+
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'vendors',
+          chunks: 'all',
+        },
+      },
+    },
+  },
+
+  plugins: [
+    new PugPlugin({
+      extractComments: false, // disable extracting license into separate file *.LICENSE.txt
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+    ],
+  },
+};