+++ test/cases/resolve-assets-split-chunk/src/assets/js/module-b.js
@@ -1,6 +1,0 @@
-const lib = require('./lib');
-const ansis = require('ansis');
-let value = lib.methodB();
-value = ansis.strip(value);
-
-module.exports = value;