+++ test/cases/split-chunk-node-module-source/src/main.js
@@ -1,0 +1,3 @@
+import 'bootstrap';
+
+console.log('>> main.js');