+++ test/cases/split-chunk-node-module-source/webpack.config.js
@@ -1,0 +1,57 @@
+const path = require('path');
+const PugPlugin = require('../../../');
+
+module.exports = {
+  mode: 'production',
+  //mode: 'development',
+
+  entry: {
+    index: './src/index.pug',
+  },
+
+  output: {
+    path: path.join(__dirname, 'public/'),
+    publicPath: '/',
+    filename: 'assets/js/[name].[contenthash:8].js',
+    clean: true,
+  },
+
+  optimization: {
+    runtimeChunk: 'single',
+    splitChunks: {
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          //test: /[\\/]node_modules[\\/].+\.(js|ts)$/, // use it when in Pug is defined CSS from node modules
+          name: 'vendors',
+          chunks: 'all',
+        },
+      },
+    },
+  },
+
+  plugins: [
+    new PugPlugin({
+      //pretty: true,
+      extractCss: {
+        filename: 'assets/css/[name].[contenthash:8].css',
+      },
+    }),
+  ],
+
+  module: {
+    rules: [
+      {
+        test: /\.pug$/,
+        loader: PugPlugin.loader,
+        options: {
+          method: 'render',
+        },
+      },
+      {
+        test: /\.(css|sass|scss)$/,
+        use: ['css-loader', 'sass-loader'],
+      },
+    ],
+  },
+};