+++ test/cases/split-chunk-resolve-assets/expected/assets/js/416.e59d09fc.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[416],{416:(e,s,p)=>{const t=p(998),h=p(773);let k=t.methodA();k=h.strip(k),e.exports=k}}]);