+++ test/cases/split-chunk-resolve-assets/expected/assets/js/541.25e78808.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[541,168],{541:(o,e,l)=>{const s=l(416),c=l(691);console.log("modA: ",s),console.log("modB: ",c)}}]);