+++ test/cases/split-chunk-resolve-assets/expected/assets/js/703.084785d8.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[703],{703:(e,s,h)=>{const k=h(998).methodC();e.exports=k}}]);