+++ test/cases/split-chunk-resolve-assets/expected/assets/js/870.575e589b.js
@@ -1,0 +1,1 @@
+(self.webpackChunk=self.webpackChunk||[]).push([[870,17],{870:(o,e,l)=>{const s=l(691),c=l(703);console.log("modB: ",s),console.log("modC: ",c)}}]);