+++ test/cases/split-chunk-resolve-assets/src/assets/js/lib.js
@@ -1,0 +1,13 @@
+const Lib = {
+  methodA() {
+    return 'A';
+  },
+  methodB() {
+    return 'B';
+  },
+  methodC() {
+    return 'C';
+  },
+};
+
+module.exports = Lib;