+++ test/cases/split-chunk-resolve-assets/src/assets/js/main-a.js
@@ -1,0 +1,5 @@
+const modA = require('./module-a');
+const modB = require('./module-b');
+
+console.log('modA: ', modA);
+console.log('modB: ', modB);