+++ test/cases/split-chunk-resolve-assets/src/assets/js/main-b.js
@@ -1,0 +1,5 @@
+const modB = require('./module-b');
+const modC = require('./module-c');
+
+console.log('modB: ', modB);
+console.log('modC: ', modC);